// src/app/main/logistics/material-request-management/shared/material-request-output.interface.ts
export interface MaterialRequestOutput {
  references?: string;
  item?: string;
  recepient?: string;
  situation?: string;
  illustration?: string;
  sn?: string;
  createdBy?: string;
  createdOn?: string;
  origin?: string;
  pn?: string;
  designation?: string;
  quantity?: string;
  desiredPotential?: string;
  site?: string;
  warehouse?: string;
  desiredOn?: string;
  criticity?: string;
  status?: string;
  progress?: string;
  served?: string;
  booked?: string;
  notBooked?: string;
  potentialUnit?: string;
}